% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{adjusted}
\alias{adjusted}
\title{adjusted is a reactive function that return a list containing multiple data frames
with the adj.P.val, logFC and P.value selected for the corresponding groups}
\usage{
adjusted(csvf)
}
\arguments{
\item{csvf}{Data frame corresponding to the Alltoptable}
}
\value{
adjusted a reactive list of data frames
}
\description{
adjusted is a reactive function that return a list containing multiple data frames
with the adj.P.val, logFC and P.value selected for the corresponding groups
}
