% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{bool}
\alias{bool}
\title{bool is a reactive function that return the bool value in the local environment}
\usage{
bool(value)
}
\value{
bool a reactive boolean outside the reactive environment
}
\description{
bool is a reactive function that return the bool value in the local environment
}
