% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{boolmark}
\alias{boolmark}
\title{boolmark is a reactive function returned to the tab1.R}
\usage{
boolmark()
}
\value{
showmark a reactive value of type boolean corresponding to the loading status by default it is set to True
}
\description{
boolmark is a reactive function returned to the tab1.R
}
