% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selgroup.R
\name{boxChooser}
\alias{boxChooser}
\title{boxChooser is a shiny module which aims is to allows users to select group(s) or comparison(s)}
\usage{
boxChooser(input, output, session, label, data, group, case, empty = F)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{label}{Label associate to the widget}

\item{data}{A Reactive expression containing a character vector of each group(s) or contrast(s)}

\item{group}{A Reactive dataframe corresponding to the pData}

\item{case}{A Numeric value}

\item{empty}{A boolean value (if true all the box are selected)}
}
\value{
A reactive vector containing the selected box
}
\description{
boxChooser is a shiny module which aims is to allows users to select group(s) or comparison(s)
}
