% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmodule.R
\name{catHm}
\alias{catHm}
\title{catHm is a shiny widget which aims is to select the categories for the functional analysis}
\usage{
catHm(id, label = "Run Analysis")
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
catHm is a shiny widget which aims is to select the categories for the functional analysis
}
