% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{chartofa}
\alias{chartofa}
\title{This function returns a transformed data frame of character type to a data frame of factor type}
\usage{
chartofa(datach)
}
\arguments{
\item{datach}{a data frames}
}
\value{
\datach a data frame
}
\description{
This function returns a transformed data frame of character type to a data frame of factor type
}
