% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selgroup.R
\name{checkboxElements}
\alias{checkboxElements}
\title{checkboxElements is a global namespace containing a list of shiny widgets to check group or comparison(s)}
\usage{
checkboxElements(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widgets in the gui
}
\description{
checkboxElements is a global namespace containing a list of shiny widgets to check group or comparison(s)
}
