% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{choix_cont}
\alias{choix_cont}
\title{choix_cont is a reactive function that return the contrast selected by the user}
\usage{
choix_cont(input)
}
\arguments{
\item{input$cont}{a set of contrasts selected by the user}
}
\value{
choix_cont a set of characters input
}
\description{
choix_cont is a reactive function that return the contrast selected by the user
}
