% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{choix_grp}
\alias{choix_grp}
\title{choix_test is a reactive function in the aim of selecting different groups}
\usage{
choix_grp(indiv)
}
\arguments{
\item{indiv}{input id corresponding to the checkboxgroup for the different groups}
}
\value{
a reactive value of type character for the different groups selected
}
\description{
choix_test is a reactive function in the aim of selecting different groups
}
