% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{choix_grpca}
\alias{choix_grpca}
\title{choix_grpca is a reactive function in the aim of selecting different groups}
\usage{
choix_grpca(indivpca)
}
\arguments{
\item{indivpca}{a vector input corresponding to the selected groups}
}
\value{
a reactive value of type character for the different groups selected
}
\description{
choix_grpca is a reactive function in the aim of selecting different groups
}
