% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{clustegrep}
\alias{clustegrep}
\title{clustergrep is a reactive function which aim is to return a list of genes for the selected cluster without the non-annotated genes}
\usage{
clustegrep(hm, cutgo)
}
\arguments{
\item{hm}{data frame of the significant genes associated with the corresponding cluster index}

\item{cutgo}{a numeric input}
}
\value{
list of genes
}
\description{
clustergrep is a reactive function which aim is to return a list of genes for the selected cluster without the non-annotated genes
}
