% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{col_choice1}
\alias{col_choice1}
\title{col_choice1 is a reactive function that return a character color}
\usage{
col_choice1(col1)
}
\arguments{
\item{col1}{input character color for the highest values}
}
\value{
col_choice1  a reactive value
}
\description{
col_choice1 is a reactive function that return a character color
}
