% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{col_choice3}
\alias{col_choice3}
\title{col_choice3 is a reactive function that return a character color}
\usage{
col_choice3(input)
}
\arguments{
\item{input}{col3 a character color for the lowest values}
}
\value{
col_choice3 reactive value
}
\description{
col_choice3 is a reactive function that return a character color
}
