% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{colname}
\alias{colname}
\title{colname is a reactive function which aim is to show or hide the colnames}
\usage{
colname(input)
}
\arguments{
\item{input$colname}{a boolean radio button input}
}
\value{
colname a reactive boolean value
}
\description{
colname is a reactive function which aim is to show or hide the colnames
}
