% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcol.R
\name{colorChooser}
\alias{colorChooser}
\title{colorChooser is a shiny module which aims is to dynamically affect for each group a color picker}
\usage{
colorChooser(input, output, session, data)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{data}{A reactive character vector for each selected group}
}
\value{
A reactive character vector that associated for each group a selected color
}
\description{
colorChooser is a shiny module which aims is to dynamically affect for each group a color picker
}
