% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{colorfluidhm}
\alias{colorfluidhm}
\title{colorfluidhm is a reactive function wich aim is to group colors side by side
depending of the number of groups odd or even for  the gui.}
\usage{
colorfluidhm(cols)
}
\arguments{
\item{cols}{a reactive number of widget-s}
}
\value{
html code interpreted by shiny
}
\description{
colorfluidhm is a reactive function wich aim is to group colors side by side
depending of the number of groups odd or even for  the gui.
}
