% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{colorfluidpca}
\alias{colorfluidpca}
\title{colorfluidpca is a reactive function wich aim is to group colors side by side
depending of the number of groups odd or even for  the gui.}
\usage{
colorfluidpca(colspca)
}
\arguments{
\item{colspca}{a reactive number of widget-s}
}
\value{
html code interpreted by shiny
}
\description{
colorfluidpca is a reactive function wich aim is to group colors side by side
depending of the number of groups odd or even for  the gui.
}
