% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{colors}
\alias{colors}
\title{colors is a reactive function which aim is to create as many variables as groups}
\usage{
colors(mycolgrp)
}
\arguments{
\item{mycolgrp}{a reactive data frame}
}
\value{
colors a reactive  list containing the different variable names
}
\description{
colors is a reactive function which aim is to create as many variables as groups
}
