% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{colorspca}
\alias{colorspca}
\title{colorspca is a reactive function which aim is to create as many variables as groups}
\usage{
colorspca(mycolgrppca)
}
\arguments{
\item{mycolgrppca}{a reactive data frame}
}
\value{
colorspca a reactive  list containing the different variable names
}
\description{
colorspca is a reactive function which aim is to create as many variables as groups
}
