% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{cols}
\alias{cols}
\title{cols is a reactive function which aim is to dynamically create widgets in function of the number of groups}
\usage{
cols(palette, mycolgrp, mypaletA)
}
\arguments{
\item{palette}{a local list defined in the environment}

\item{mycolgrp}{a dataframe representing the selected groups}
}
\value{
cols a reactive number of widget-s
}
\description{
cols is a reactive function which aim is to dynamically create widgets in function of the number of groups
}
