% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{colspca}
\alias{colspca}
\title{colspca is a reactive function which aim is to dynamically create widgets depending on the number of groups}
\usage{
colspca(brewer.pal, mycolgrppca)
}
\arguments{
\item{brewer.pal}{a local list defined in the RcolorBrewer package}

\item{mycolgrppca}{a dataframe representing the selected groups}
}
\value{
colspca a reactive number of widget-s
}
\description{
colspca is a reactive function which aim is to dynamically create widgets depending on the number of groups
}
