% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{csvDecimal}
\alias{csvDecimal}
\title{csvDecimal is a shiny widget to select the decimal separator}
\usage{
csvDecimal(id, label = "Decimal")
}
\arguments{
\item{id}{Shiny id}

\item{label}{Shiny label}
}
\value{
Widget in the gui
}
\description{
csvDecimal is a shiny widget to select the decimal separator
}
