% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{csvf}
\alias{csvf}
\title{boolmark is a reactive function returned to the tab1.R}
\usage{
csvf(file)
}
\arguments{
\item{file}{html id for files loaded in csv format}
}
\value{
showmark a reactive value of type boolean corresponding to the loading status by default it is set to True

csvf a reactive value of type list containing three data frames toptable, workingset and the pData

list of data frame objects

a reactive value of type boolean set to False
}
\description{
boolmark is a reactive function returned to the tab1.R
}
