% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{csvFile}
\alias{csvFile}
\title{csvFile is shiny module which aims is to alert users on the status of his upload and return an ordered list of dataframe}
\usage{
csvFile(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{stringsAsFactors}{}
}
\value{
The reactive that yields a list of dataframes

a reactive value of type boolean corresponding to the loading status by default it is set to True

Reactive function returned to the tab1.R

\showmark a reactive value of type boolean set to False
}
\description{
csvFile is shiny module which aims is to alert users on the status of his upload and return an ordered list of dataframe
}
