% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{csvFileInput}
\alias{csvFileInput}
\title{### Author: Franck Soubès
### Bioinformatics Master Degree - University of Bordeaux, France
### Link: https://github.com/fsoubes/MA_Trix_App
### Where: GET-TRiX's facility
### Application: MATRiX is a shiny application for Microarray Analysis on Transcriptomic impact of Xenobiotics
### Licence: GPL-3.0
csvFileInput is a shiny widget for uploading data locally}
\usage{
csvFileInput(id, label = "CSV file", multiple = T)
}
\arguments{
\item{id}{Shiny id}

\item{label}{Shiny label}

\item{multiple}{Shiny boolean for selecting or not multiple files in the browser}
}
\value{
Widget in the gui
}
\description{
### Author: Franck Soubès
### Bioinformatics Master Degree - University of Bordeaux, France
### Link: https://github.com/fsoubes/MA_Trix_App
### Where: GET-TRiX's facility
### Application: MATRiX is a shiny application for Microarray Analysis on Transcriptomic impact of Xenobiotics
### Licence: GPL-3.0
csvFileInput is a shiny widget for uploading data locally
}
