% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{csvIdentifier}
\alias{csvIdentifier}
\title{csvIdentifier is a shiny widget for loading data from the server}
\usage{
csvIdentifier(id, label = "Unique Identifier")
}
\arguments{
\item{id}{Shiny id}

\item{label}{Shiny label}
}
\value{
Widget in the gui
}
\description{
csvIdentifier is a shiny widget for loading data from the server
}
