% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{cutfinal}
\alias{cutfinal}
\title{cutfinal is a reactive function that return heatmap or ggplot2 object}
\usage{
cutfinal(p, input, new_data, rownamtoX, groups, input2, input3)
}
\arguments{
\item{p}{heatmap object}

\item{input}{a numeric value to cut the dendogram}

\item{new_data}{a data frame with specific columns depending on the user's choices}

\item{rownamtoX}{a data frame}

\item{groups}{a data frame of the choosen groups}

\item{input2}{heatmap object}

\item{input3}{a character to select the plot to display heatmap, boxplot or stripchart}
}
\value{
\cutfinal a ggplot object or heatmapply object
}
\description{
cutfinal is a reactive function that return heatmap or ggplot2 object
}
