% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutheat.R
\name{cutHeatmaps}
\alias{cutHeatmaps}
\title{cutHeatmaps if a function that takes as input an heatmap object and depending on the cut height and the cluster
choosen render a ggplot object or an heatmap object}
\usage{
cutHeatmaps(hmp, height, exprData, groups, cexcol = 1, cexrow = 1,
  labrow = F, genename = NULL, fileType = "png", scales = "row",
  meanGrp = F, mypal = NULL, type = "None", las = 2,
  distfun = "cor", palette.col = NULL, num = 4, ...)
}
\arguments{
\item{hmp}{An heatmap object}

\item{height}{A numeric value to cut the dendogram}

\item{exprData}{A data frame with specific columns depending on the user's choices}

\item{groups}{A data frame of the choosen groups}

\item{cexcol}{A positive numbers, used as cex.axis in for the row or column axis labeling}

\item{cexrow}{A positive numbers, used as cex.axis in for the row or column axis labeling}

\item{labrow}{A character vectors with row and column labels to use}

\item{genename}{A character vector of gene symbols}

\item{fileType}{A character to select the plot to display heatmap, boxplot or stripchart}

\item{scales}{A character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none"}

\item{meanGrp}{A boolean value to computes the mean for each groups; default = F}

\item{type}{A character to select the plot to display heatmap, boxplot or stripchart}

\item{las}{A numeric value}

\item{distfun}{Function used to compute the distance (dissimilarity) between both rows and columns.}

\item{palette.col}{A character vector of colors}

\item{num}{An item of the heatmap object corresponding to a specific cluster choosen by the user}

\item{...}{}
}
\value{
a ggplot object or heatmapply object
}
\description{
cutHeatmaps if a function that takes as input an heatmap object and depending on the cut height and the cluster
choosen render a ggplot object or an heatmap object
}
