% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdegenes.R
\name{cutoffElements}
\alias{cutoffElements}
\title{cutoffElements is a global namespace containing a list of shiny widgets filter the restable with cutoff parameters}
\usage{
cutoffElements(id, c1 = 6, c2 = 6)
}
\arguments{
\item{id}{Shiny id}

\item{c1}{A numeric value to specify the column size of the first element}

\item{c2}{A numeric value to specify the column size of the second element}
}
\value{
Widgets in the gui
}
\description{
cutoffElements is a global namespace containing a list of shiny widgets filter the restable with cutoff parameters
}
