% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{cuttfinal}
\alias{cuttfinal}
\title{cutfinal is a reactive function that return heatmap or ggplot2 object}
\usage{
cuttfinal(hmobj, cut, new_data, rownamtoX, groups, cutcluster, cutinfo)
}
\arguments{
\item{cut}{a numeric input corresponding to the height where the dendogram is cut}

\item{new_data}{a data frame with specific columns depending on the user's choices}

\item{rownamtoX}{a data frame}

\item{groups}{a data frame of the choosen groups}

\item{cutcluster}{a numeric input corresponding to the selected cluster to display}

\item{cutinfo}{a character input to select the plot to display heatmap, boxplot or stripchart}

\item{hmobj$obj}{heatmap object}
}
\value{
a ggplot object or heatmapply object
}
\description{
cutfinal is a reactive function that return heatmap or ggplot2 object
}
