% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{data_sign}
\alias{data_sign}
\title{Reactive function that return a comparison data frame with the specific user's selection
it's an old function}
\usage{
data_sign(adjusted)
}
\arguments{
\item{adjusted}{a reactive list of data frames}
}
\value{
new_data a  data frame with all the individuals selected
}
\description{
Reactive function that return a comparison data frame with the specific user's selection
it's an old function
}
