% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{davidquery}
\alias{davidquery}
\title{davidquery is a function which aim is to querrying DWS and performing go term enrichment analysis}
\usage{
davidquery(entrezids, species, mycat)
}
\arguments{
\item{entrezids}{list of entrez IDS}

\item{species}{character name of the species whose genes are enriched}

\item{mycat}{category of the enrichment analysis: MF, CC, BP or KEGG pathway}
}
\value{
list of data frames for each cluster containing
}
\description{
davidquery is a function which aim is to querrying DWS and performing go term enrichment analysis
}
