% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{davidqueryvenn}
\alias{davidqueryvenn}
\title{davidqueryvenn is a function which aim to query DWS with a list of entrez ids to generate the Term/Gene cluster report}
\usage{
davidqueryvenn(entrezids, species)
}
\arguments{
\item{entrezids}{List of entrez IDS}

\item{species}{character name of the species whose genes are enriched}
}
\value{
david object of S4 class
}
\description{
davidqueryvenn is a function which aim to query DWS with a list of entrez ids to generate the Term/Gene cluster report
}
