% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{davidurl}
\alias{davidurl}
\title{davidurl is a reactive function that aim is to return an url of grouped genes}
\usage{
davidurl(clustergrep)
}
\arguments{
\item{clustergrep}{list of genes}
}
\value{
character
}
\description{
davidurl is a reactive function that aim is to return an url of grouped genes
}
