% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{davidwebservice}
\alias{davidwebservice}
\title{davidwebservice is an eventreactive function which aim is to querrying the DWS to return a dataframe summary}
\usage{
davidwebservice(GO, hm, Species, catinfo)
}
\arguments{
\item{GO}{clickable event}

\item{hm}{data frame of the significant genes associated with the corresponding cluster index}

\item{Species}{list of annotated elements}

\item{catinfo}{vector of enrichment categories, BP, CC, MF, Kegg}
}
\value{
data frame
}
\description{
davidwebservice is an eventreactive function which aim is to querrying the DWS to return a dataframe summary
}
