% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartconverter.R
\name{DftoHighjson}
\alias{DftoHighjson}
\title{DftoHighjson is a function which aims is to reshape a DAVID restable dataframe to a json format}
\usage{
DftoHighjson(data, method)
}
\arguments{
\item{data}{A Functional Annotation Summary dataframe (DAVID output)}

\item{method}{A character to specify if users want to display top term based on the pvalue or the fold Enrichment}
}
\value{
Json object to be pass to highchart library
}
\description{
DftoHighjson is a function which aims is to reshape a DAVID restable dataframe to a json format
}
