% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{dirModuleUI}
\alias{dirModuleUI}
\title{dirModuleUI is a shiny widget for loading data from the server}
\usage{
dirModuleUI(id, multiple = T)
}
\arguments{
\item{id}{Shiny id}

\item{multiple}{Shiny boolean for selecting or not multiple files in the browser}
}
\value{
Widget in the gui
}
\description{
dirModuleUI is a shiny widget for loading data from the server
}
