% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/heatmtruncated.R
\name{distcor}
\alias{distcor}
\alias{distcor}
\title{distcor is a function that computes the distance correlation}
\usage{
distcor(x)

distcor(x)
}
\arguments{
\item{x}{}

\item{x}{}
}
\value{
a matrix distance

a matrix distance
}
\description{
distcor is a function that computes the distance correlation

distcor is a function that computes the pearson correlation
}
