% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{disteucl}
\alias{disteucl}
\title{disteuc is a function that computes the euclidean correlation}
\usage{
disteucl(x)
}
\arguments{
\item{x}{}
}
\value{
a matrix distance
}
\description{
disteuc is a function that computes the euclidean correlation
}
