% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{distman}
\alias{distman}
\title{disteuc is a function that computes the euclidean distance matrices}
\usage{
distman(x)
}
\arguments{
\item{x}{A vector}
}
\value{
a matrix distance
}
\description{
disteuc is a function that computes the euclidean distance matrices
}
