% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savemodules.R
\name{downloadFiles}
\alias{downloadFiles}
\title{downloadFiles is a shiny widget which aims is to download a reactive plot}
\usage{
downloadFiles(id, label = "Save your Scree plot")
}
\arguments{
\item{id}{Shiny id}

\item{label}{Shiny label}
}
\value{
Widget in the gui
}
\description{
downloadFiles is a shiny widget which aims is to download a reactive plot
}
