% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Savemoduletable.R
\name{downloadFilestab}
\alias{downloadFilestab}
\title{downloadFilestab is a shiny widget which aims is to download a reactive plot}
\usage{
downloadFilestab(id, label = "Save your Scree plot")
}
\arguments{
\item{id}{Shiny id}

\item{label}{Shiny label}
}
\value{
Widget in the gui
}
\description{
downloadFilestab is a shiny widget which aims is to download a reactive plot
}
