% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Savemoduletable.R
\name{downoutputables}
\alias{downoutputables}
\title{downoutputables is a shiny module which aims is to export reactive table}
\usage{
downoutputables(input, output, session, projectname, suffix = "plot.csv",
  data = NULL, cont = NULL, xlsx = F, case)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{projectname}{A reactive character (MA project or session data)}

\item{suffix}{A character}

\item{data}{A reactive expression to be plot}

\item{cont}{A reactive character vector of the selected comparison(s)}

\item{xlsx}{A boolean value}

\item{case}{A numeric value}
}
\description{
downoutputables is a shiny module which aims is to export reactive table
}
