% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnhancedVolcano.R
\name{EnhancedVolcano}
\alias{EnhancedVolcano}
\title{Volcano plots represent a useful way to visualise the results of differential expression analyses. Here, we present a highly-configurable function that
produces publication-ready volcano plots. EnhancedVolcano will attempt to fit as many transcript names in the plot window as possible,
thus avoiding 'clogging' up the plot with labels that could not otherwise have been read.}
\usage{
EnhancedVolcano(toptable, lab, x, y, selectLab = NULL,
  displaylab = NULL, findfamily = NULL, topgenes = NULL,
  regulationvolc = NULL, xlim = c(min(toptable[, x], na.rm = TRUE),
  max(toptable[, x], na.rm = TRUE)), ylim = c(0, max(-log10(toptable[,
  y]), na.rm = TRUE) + 5), xlab = bquote(~Log[2] ~ "fold change"),
  ylab = bquote(~-Log[10] ~ italic(P)), axisLabSize = 16,
  pCutoff = 0.05, pLabellingCutoff = pCutoff, FCcutoff = 2,
  title = "", titleLabSize = 16, transcriptPointSize = 2.8,
  transcriptLabSize = 5, col = c("grey30", "forestgreen", "royalblue",
  "red2"), colAlpha = 1/2, legend = c("NS", "Log2 FC", "P",
  "P & Log2 FC"), legendPosition = "top", legendLabSize = 10,
  legendIconSize = 3, DrawConnectors = FALSE, widthConnectors = 0.5,
  colConnectors = "black", cutoffLineType = "longdash",
  cutoffLineCol = "black", cutoffLineWidth = 0.4)
}
\arguments{
\item{toptable}{Requires at least the following: column for transcript names (can be rownames); a column for log2 fold changes; a column for nominal or adjusted p-value. REQUIRED.}

\item{lab}{A column name in toptable containing transcript names. Can be rownames(toptable). REQUIRED.}

\item{x}{A column name in toptable containing log2 fold changes. REQUIRED.}

\item{y}{A column name in toptable containing nominal or adjusted p-values. REQUIRED.}

\item{selectLab}{A vector containing a subset of lab. DEFAULT = NULL. OPTIONAL.}

\item{displaylab}{A comma-separated values corresponding to the displayed genes. DEFAULT = NULL. OPTIONAL.}

\item{findfamily}{A character to parse familiy of genes. DEFAULT = NULL. OPTIONAL.}

\item{topgenes}{A numeric value to display the top n genes based on the regulation.DEFAULT = NULL. OPTIONAL.}

\item{regulationvolc}{A character for the regulation ("up", "down", "both") .DEFAULT = NULL. OPTIONAL.}

\item{xlim}{Limits of the x-axis. DEFAULT = c(min(toptable[,x], na.rm=TRUE),max(toptable[,x], na.rm=TRUE)). OPTIONAL.}

\item{ylim}{Limits of the y-axis. DEFAULT = c(0, max(-log10(toptable[,y]), na.rm=TRUE) + 5). OPTIONAL.}

\item{xlab}{Label for x-axis. DEFAULT = bquote(Log[2]~ "fold change"). OPTIONAL.}

\item{ylab}{Label for y-axis. DEFAULT = bquote(-Log[10]~ P)).OPTIONAL.}

\item{axisLabSize}{Size of x- and y-axis labels. DEFAULT = 18. OPTIONAL.}

\item{pCutoff}{Cut-off for statistical significance. A horizontal line will be drawn at -log10(pCutoff). DEFAULT = 10e-6. OPTIONAL.}

\item{pLabellingCutoff}{Labelling cut-off for statistical significance. DEFAULT = pCutoff. OPTIONAL}

\item{FCcutoff}{Cut-off for absolute log2 fold-change. Vertical lines will be drawn at the negative and positive values of log2FCcutoff. DEFAULT =1.0. OPTIONAL.}

\item{title}{Plot title. DEFAULT = 'Volcano plot'. OPTIONAL.}

\item{titleLabSize}{Plot subtitle. DEFAULT = 'Bioconductor package, EnhancedVolcano'. OPTIONAL.}

\item{transcriptPointSize}{Size of plotted points for each transcript. DEFAULT = 0.8. OPTIONAL.}

\item{transcriptLabSize}{Size of labels for each transcript. DEFAULT = 3.0. OPTIONAL.}

\item{col}{Colour shading for plotted points, corresponding to< abs(FCcutoff) && > pCutoff, > abs(FCcutoff), < pCutoff,> abs(FCcutoff) && < pCutoff. DEFAULT = c("grey30", "forestgreen","royalblue", "red2"). OPTIONAL.}

\item{colAlpha}{Alpha for purposes of controlling colour transparency oftranscript points. DEFAULT = 1/2. OPTIONAL.}

\item{legend}{Plot legend text. DEFAULT = c("NS", "Log2 FC", "P","P & Log2 FC"). OPTIONAL.}

\item{legendPosition}{Position of legend ("top", "bottom", "left","right"). DEFAULT = "top". OPTIONAL.}

\item{legendLabSize}{Size of plot legend text. DEFAULT = 14. OPTIONAL.}

\item{legendIconSize}{Size of plot legend icons / symbols. DEFAULT = 4.0.OPTIONAL.}

\item{DrawConnectors}{Logical, indicating whether or not to connect plotlabels to their corresponding points by line connectors. DEFAULT = FALSE.OPTIONAL.}

\item{widthConnectors}{Line width of connectors. DEFAULT = 0.5. OPTIONAL.}

\item{colConnectors}{Line colour of connectors. DEFAULT = 'grey10'. OPTIONAL.}

\item{cutoffLineType}{Line type for FCcutoff and pCutoff ("blank","solid", "dashed", "dotted", "dotdash", "longdash", "twodash").DEFAULT = "longdash". OPTIONAL.}

\item{cutoffLineCol}{Line colour for FCcutoff and pCutoff. DEFAULT ="black". OPTIONAL.}

\item{cutoffLineWidth}{Line width for FCcutoff and pCutoff. DEFAULT = 0.4. OPTIONAL.}
}
\value{
A list of two elements
}
\description{
Volcano plots represent a useful way to visualise the results of differential expression analyses. Here, we present a highly-configurable function that
produces publication-ready volcano plots. EnhancedVolcano will attempt to fit as many transcript names in the plot window as possible,
thus avoiding 'clogging' up the plot with labels that could not otherwise have been read.
}
\author{
Kevin Blighe <kevin@clinicalbioinformatics.co.uk>
}
