% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{evaluatesign}
\alias{evaluatesign}
\title{This function returns an integer for the number of significant genes}
\usage{
evaluatesign(adj, elem, pv)
}
\arguments{
\item{adj}{a data frame}

\item{elem}{a list}

\item{pv}{a list}
}
\value{
\grp1 of class data frame
}
\description{
This function returns an integer for the number of significant genes
}
