% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{filterjvenn}
\alias{filterjvenn}
\title{filterjvenn is a function which takes as input a list of genes or probes obtained by selecting a set in the venn diagram and return
the association of this list with the logFC}
\usage{
filterjvenn(jvennlist, selcontjv, restab, idcol, usersel,
  venngeneslist = NULL)
}
\arguments{
\item{jvennlist}{A vector of genes generated from the jvenn}

\item{selcontjv}{A character vector of the selected contrast(s)}

\item{restab}{A dataframe corresponding to the statistical table}

\item{idcol}{A character value (transcripts/probes)}

\item{usersel}{A character value (display genes or probes with the venn diagram)}

\item{venngeneslist}{A list of transcripts/probes generated from the Vennlist function which aims is to remove probes:transcripts that are not significant}
}
\value{
a list of genes associated with the logFC
}
\description{
filterjvenn is a function which takes as input a list of genes or probes obtained by selecting a set in the venn diagram and return
the association of this list with the logFC
}
