% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartconverter.R
\name{formatCategory}
\alias{formatCategory}
\title{formatCategory is a function which aims is to return a json structure}
\usage{
formatCategory(dframeCategory)
}
\arguments{
\item{dframeCategory}{A dataframe to be processed}
}
\value{
Json structure
}
\description{
formatCategory is a function which aims is to return a json structure
}
