% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{formated}
\alias{formated}
\title{formated is a reactive function that return the indexes for the signficant genes}
\usage{
formated(user_group, fc, method2, pval, maxgen)
}
\arguments{
\item{user_group}{a list of three data frame with rows selected according to the contrasts selected}

\item{intput$fc}{a numeric FC selected}

\item{input$method2}{a character method, default = BH}

\item{input$pval}{a numeric pvalue}

\item{input$maxgen}{a numeric maxgen, default = NULL}
}
\value{
formated a reactive data frame with the indexes corresponding to the sigificant genes
}
\description{
formated is a reactive function that return the indexes for the signficant genes
}
