% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{formating}
\alias{formating}
\title{formating is an alpha version of the higher elaborate decTestTRiX function that return a vector list of the signficant genes
for a certain treshold}
\usage{
formating(adj, pval)
}
\arguments{
\item{pval}{}
}
\description{
formating is an alpha version of the higher elaborate decTestTRiX function that return a vector list of the signficant genes
for a certain treshold
}
