% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartconverter.R
\name{formatPoint}
\alias{formatPoint}
\title{formatPoint is a function which aims is to reutnr a list of information for each category (top 10)}
\usage{
formatPoint(index)
}
\arguments{
\item{index}{A dataframe to be processed}
}
\value{
A list of usefull information from the david restable
}
\description{
formatPoint is a function which aims is to reutnr a list of information for each category (top 10)
}
