% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdegenes.R
\name{getDegenes}
\alias{getDegenes}
\title{getDegenes is a shiny module which aims is to return an index or list of each significant genes from the restable for a set of comparison(s)}
\usage{
getDegenes(input, output, session, data, dflogfc = NULL, maxDe = NULL,
  case = NULL, reg = NULL, meth = NULL)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{data}{A subset data frame with the selected comparisons for the adj.p.val or p.val and with the logfc for the hm}

\item{dflogfc}{A subset data frame with the selected comparisons for the logfc}

\item{maxDe}{A reactive numeric value}

\item{case}{A numeric value (case = 1 for heatmap and strip and 2 for venn diagram)}

\item{reg}{A reactive character}

\item{meth}{A reactive Character}
}
\value{

}
\description{
getDegenes is a shiny module which aims is to return an index or list of each significant genes from the restable for a set of comparison(s)
}
