% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstrip_groups.r
\name{ggstrip_groups}
\alias{ggstrip_groups}
\title{ggstrip_groups which aims is to return a strip plot of the expression for a selected gene for each groups.}
\usage{
ggstrip_groups(grps, wSet, probesID)
}
\arguments{
\item{grps}{A character vector sorted by group of each sample}

\item{wSet}{A dataframe of the normalised expression}

\item{probesID}{A vector of the expression of a selected row from the outputted table}
}
\value{
a ggplot object
}
\description{
ggstrip_groups which aims is to return a strip plot of the expression for a selected gene for each groups.
}
