% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{global}
\alias{global}
\title{global is a ReactiveValues function}
\usage{
global()
}
\arguments{
\item{clicked}{bool set to FALSE}

\item{heatm}{input triggered by click event}
}
\value{
clicked a boolean which can be TRUE or FALSE
}
\description{
global is a ReactiveValues function
}
