% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{gosearch}
\alias{gosearch}
\title{gosearch is a function that return a list of data frame containing the go ids for the different clusters}
\usage{
gosearch(hm01, species, ids, clusterlist)
}
\arguments{
\item{hm01}{A Data frame object}

\item{species}{A Character vector}

\item{ids}{Package use to perform the enrichment}

\item{clusterlist}{List}
}
\value{
list of data frames
}
\description{
gosearch is a function that return a list of data frame containing the go ids for the different clusters
}
