% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/heatmtruncated.R
\name{hclustfun}
\alias{hclustfun}
\alias{hclustfun}
\title{hclustfun is a function that performs a hiearchical cluster analysis with the method of Ward}
\usage{
hclustfun(d)

hclustfun(d)
}
\arguments{
\item{d}{}

\item{d}{}
}
\value{
an object of class hclust

an object of class hclust
}
\description{
hclustfun is a function that performs a hiearchical cluster analysis with the method of Ward

hclustfun is a function that performs a hiearchical cluster analysis with the method of Ward
}
