% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{heatmtoclust}
\alias{heatmtoclust}
\title{This function returns a data frame of the significant genes associated with the corresponding cluster index}
\usage{
heatmtoclust(hmp01_All, exprData, pval, height = 5)
}
\arguments{
\item{hmp01_All}{An heatmap object}

\item{exprData}{An index vector of significant genes from the restable}

\item{pval}{A data frame of the restable}

\item{height}{A cutoff point of a dendogram}
}
\value{
A data frame (unique ids, gene symbols, pvalue and cluster)
}
\description{
This function returns a data frame of the significant genes associated with the corresponding cluster index
}
