% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchartconverter.R
\name{highchartsConvert}
\alias{highchartsConvert}
\title{highchartsConvert is a function which aims is to convert a dataframe object to the json format
by spliting the dataframe by categories to be passed to the highchart library}
\usage{
highchartsConvert(DftoHighjson)
}
\arguments{
\item{DftoHighjson}{A dataframe to be processed}
}
\value{
A json object example:{"GOTERM_BP_ALL" : { "x" : 0.005 , "y" : 2 , "Term" : "inactivation of MAPK activity" , "Go" : "GO:0000188"}, ..., "GOTERM_CC_ALL : { ...}}
}
\description{
highchartsConvert is a function which aims is to convert a dataframe object to the json format
by spliting the dataframe by categories to be passed to the highchart library
}
