% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{indnull}
\alias{indnull}
\title{indnull is a reactive function that return a vector for the contrasts with 0 genes significant at a treshold set to 5%}
\usage{
indnull(vennlistshi)
}
\arguments{
\item{vennlistshi}{a list}
}
\value{
indnull a reactive vector
}
\description{
indnull is a reactive function that return a vector for the contrasts with 0 genes significant at a treshold set to 5%
}
