% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Intersect}
\alias{Intersect}
\title{Intersect is a function that takes a list as argument and return the identical elements between those lists}
\usage{
Intersect(x)
}
\arguments{
\item{x}{A list of elements}
}
\value{
A vector
}
\description{
Intersect is a function that takes a list as argument and return the identical elements between those lists
}
\examples{
x <- c(sort(sample(1:20, 9)))
y <- c(sort(sample(3:23, 7)))
test = list()
test[[1]] = x
test[[2]] = y
Intersect(test) = 9,19
}
