% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{labeled}
\alias{labeled}
\title{labeled is a reactive function which aim is to display or not the labels in the PCA render plot}
\usage{
labeled(label)
}
\arguments{
\item{label}{a boolean input}
}
\value{
Labeled a reactive  boolean depending of the user's choice to display or not the labels
}
\description{
labeled is a reactive function which aim is to display or not the labels in the PCA render plot
}
