% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{list_ind}
\alias{list_ind}
\alias{list_ind}
\title{list_ind is a reactive function in the aim of having selected groups in a list}
\usage{
list_ind(indivpca)

list_ind(indivpca)
}
\arguments{
\item{indivpca}{specific of the individuals data frame}

\item{indiv}{input id corresponding to the checkboxgroup for the different groups}
}
\value{
a reactive list for the different individuals selected

a reactive list for the different individuals selected
}
\description{
list_ind is a reactive function in the aim of having selected groups in a list

list_ind is a reactive function in the aim of having selected groups in a list
}
