% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{mean_grp}
\alias{mean_grp}
\title{Reactive function which aim is to return the user's input}
\usage{
mean_grp(output)
}
\arguments{
\item{input$meangrp}{character depending on the user's choice to compute the mean for each different groups in the heatmap}

\item{ouput}{character depending on the user's choice}
}
\value{
a reactive value of type string/char

mean_grp is a reactive function which aim is to return the output user in order to show this input in the UI

mean_grp a reactive value of type character depending on the user's input
}
\description{
Reactive function which aim is to return the user's input
}
