% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{meanrankgenes}
\alias{meanrankgenes}
\title{meanrankgenes is a function which aims is to return a dataframe with the average logFC for each duplicate genes}
\usage{
meanrankgenes(dtsign, stat, rankcomp = NULL, multcomp,
  regulationvolc = NULL, jvenn = F)
}
\arguments{
\item{dtsign}{A statistical dataframe (ids and logFC)}

\item{stat}{A character to grep logFC or pvalue values within the statistical table}

\item{rankcomp}{A character vector of the selected contrast(s)}

\item{multcomp}{A choosen comparison to sort the genes based on the logFC}

\item{regulationvolc}{A character to specific the regulation (both, up, down) for the volcano page}

\item{jvenn}{A boolean value}
}
\value{
A dataframe
}
\description{
meanrankgenes is a function which aims is to return a dataframe with the average logFC for each duplicate genes
}
