% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{my_intermediate}
\alias{my_intermediate}
\title{my_intermediate is a reactive function that return a character color}
\usage{
my_intermediate(col_choice1, col_choice3)
}
\arguments{
\item{col_choice1}{character color for the lowest values}

\item{col_choice3}{character color for the highest values}
}
\value{
inter a reactive character intermediate color between the lowest and the highest values
}
\description{
my_intermediate is a reactive function that return a character color
}
