% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{mycolgrp}
\alias{mycolgrp}
\title{mycolgrp is a reactive function which aim is to display the number of groups selected}
\usage{
mycolgrp(new_group)
}
\arguments{
\item{new_group}{a subset data frame of the pData}
}
\value{
mycolgrp a reactive data frame
}
\description{
mycolgrp is a reactive function which aim is to display the number of groups selected
}
