% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{mycolgrppca}
\alias{mycolgrppca}
\title{mycolgrppca is a reactive function which aim is to display the total number of groups}
\usage{
mycolgrppca(csvf, new_grouppca)
}
\arguments{
\item{csvf}{a dataframe}

\item{new_grouppca}{a reactive factor with the corresponding groups selected}
}
\value{
mycolgrppca a reactive reorder dataframe
}
\description{
mycolgrppca is a reactive function which aim is to display the total number of groups
}
