% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{mygotabres}
\alias{mygotabres}
\title{mygotabres is a function which aim is to summarise the top 10 for each different cagetogies of the  the DAVID gene set enrichment analysis data table}
\usage{
mygotabres(davtab, enrichbased)
}
\arguments{
\item{davtab}{data frame}
}
\value{
list of data frames
}
\description{
mygotabres is a function which aim is to summarise the top 10 for each different cagetogies of the  the DAVID gene set enrichment analysis data table
}
