% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{myindex}
\alias{myindex}
\title{myindex is a reactive function returning the column indices for which there's more than one significant genes}
\usage{
myindex(adjusted)
}
\arguments{
\item{adjusted}{data frame corresponding to the adjusted.pval}
}
\value{
myindex a numeric vector
}
\description{
myindex is a reactive function returning the column indices for which there's more than one significant genes
}
