% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{mypacklist}
\alias{mypacklist}
\title{mypacklist is a reactive function which aim is to display the different packages used in the current session}
\usage{
mypacklist(sessionInfo)
}
\arguments{
\item{sessionInfo}{version information about R, the OS and attached or loaded packages.}
}
\value{
a data frame
}
\description{
mypacklist is a reactive function which aim is to display the different packages used in the current session
}
