% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{mypal}
\alias{mypal}
\title{mypal is a reactive function which aim is to unlist the choice of colors}
\usage{
mypal(colors)
}
\arguments{
\item{colors}{a list of input for the different user's choice}
}
\value{
mypal a reactive  that unlist the colors attributed to the different groups
}
\description{
mypal is a reactive function which aim is to unlist the choice of colors
}
