% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{mypaletA}
\alias{mypaletA}
\title{mypaletA is a reactive function which aim is to set colors if the advanced graphical settings are not displays}
\usage{
mypaletA(colors)
}
\arguments{
\item{colors}{a list of input for the different user's choice}
}
\value{
mypaletA a reactive list of colors attributed by ranking order to the different groups
}
\description{
mypaletA is a reactive function which aim is to set colors if the advanced graphical settings are not displays
}
