% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{myrendertop}
\alias{myrendertop}
\title{myrenderedtop is a reactive function which aim is to display or not the labels in the PCA render plot}
\usage{
myrendertop(csvf)
}
\arguments{
\item{csvf}{a data frame}
}
\value{
a reactive data frame
}
\description{
myrenderedtop is a reactive function which aim is to display or not the labels in the PCA render plot
}
