% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{mysetventocsv}
\alias{mysetventocsv}
\title{mysetventocsv is a function that create a csv file of the signficant genes for the different contrasts.}
\usage{
mysetventocsv(myven)
}
\arguments{
\item{myven}{A list of significant genes for each contrasts}
}
\value{
A dataframe containing all the significant genes for each contrast(s)
}
\description{
mysetventocsv is a function that create a csv file of the signficant genes for the different contrasts.
}
