% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{mytransf}
\alias{mytransf}
\title{mytransf is a reactive function which aim is to convert entrez ID to GENE  the selected rows in the output data table}
\usage{
mytransf(davidwebservice, cutgo, davidgo_rows_selected, Species)
}
\arguments{
\item{davidwebservice}{data frame}

\item{cutgo}{a numeric input}

\item{davidgo_rows_selected}{selected rows}

\item{Species}{list of annotated elements}
}
\value{
a data frame
}
\description{
mytransf is a reactive function which aim is to convert entrez ID to GENE  the selected rows in the output data table
}
