% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{myventocsv}
\alias{myventocsv}
\title{myventocsv is a function that create a csv file of the signficant genes for the different contrasts.}
\usage{
myventocsv(myven, adj)
}
\arguments{
\item{myven}{A list of significant genes for each contrasts}

\item{adj}{A dataframe}
}
\value{
A dataframe containing all the significant genes for each contrast(s)
}
\description{
myventocsv is a function that create a csv file of the signficant genes for the different contrasts.
}
