% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{new_datapca}
\alias{new_datapca}
\title{new_datapca is a reactive function that aim is to select specific individuals in the data frame}
\usage{
new_datapca(csvf)
}
\arguments{
\item{csvf}{Data frame corresponding to the Workingset}
}
\value{
new_datapca a reactive data frame
}
\description{
new_datapca is a reactive function that aim is to select specific individuals in the data frame
}
