% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{new_group}
\alias{new_group}
\alias{new_group}
\title{new_group is an eventreactive function that select specific groups in the data frame}
\usage{
new_group(csvf, choix_grp)

new_group(csvf, choix_grp)
}
\arguments{
\item{csvf}{a Data frame corresponding to the pData table}

\item{choix_grp}{a reactive value of type character for the different groups selected}

\item{csvf}{a data frame of the pData}

\item{choix_grp}{a vector character corresponding to the defined groups set by the user}
}
\value{
new_group an eventreactive factor with the corresponding groups selected

new_group a reactive new factor with the corresponding groups
}
\description{
new_group is an eventreactive function that select specific groups in the data frame

new group is a reactive function that select specific groups in the data frame
}
