% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{new_grouppca}
\alias{new_grouppca}
\title{new_grouppca is a reactive function that select specific groups in the data frame}
\usage{
new_grouppca(heatm, csvf)
}
\arguments{
\item{heatm}{a clickable input button}

\item{csvf}{a Data frame corresponding to the pData table}
}
\value{
new_grouppca a reactive factor with the corresponding groups selected
}
\description{
new_grouppca is a reactive function that select specific groups in the data frame
}
