% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{num2cols}
\alias{num2cols}
\title{num2cols is a function that gives a character vector of color names for a given numeric vector}
\usage{
num2cols(numVector, colp = palette())
}
\arguments{
\item{numVector}{A numeric vector}

\item{colp}{A character vector}
}
\value{
An object matrix
}
\description{
num2cols is a function that gives a character vector of color names for a given numeric vector
}
