% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvmodules.R
\name{ordinput}
\alias{ordinput}
\title{ordinput is a function which aims to reorder an input list of dataframes}
\usage{
ordinput(csvnord, identifier, dec)
}
\arguments{
\item{csvnord}{A list of dataframe}

\item{identifier}{Character identifier (probes/transcripts)}

\item{dec}{Character to select the decimal separator}
}
\value{
An ordered list (WorkingSet, pData, restable)

\csv a data frame object
}
\description{
ordinput is a function which aims to reorder an input list of dataframes
}
