% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{p}
\alias{p}
\title{p is a reactive function that return a heatmap gplots object}
\usage{
p(updateheatm, hmobj)
}
\arguments{
\item{updateheatm}{clickable input button}

\item{hmobj$obj}{a reactive value object}
}
\value{
p an heatmap object isolate
}
\description{
p is a reactive function that return a heatmap gplots object
}
