% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCAplot}
\alias{PCAplot}
\title{PCAplot is a function that return a factoextra object of PCA type}
\usage{
PCAplot(PCAres, myax = c(1, 2), elips = T, rep = T,
  mylevel = groups$Grp, mylabsize = 4, dispelip = 0.8,
  labeled = "all", pal = brewer.pal(8, "Dark2"))
}
\arguments{
\item{PCAres}{A data frame with PCA attributes}

\item{myax}{A numeric vector of length 2 specifying the dimensions to be plotted}

\item{elips}{A boolean value to add ellipse to the data distribution  for the different groups; default = False}

\item{rep}{A boolean value to avoid overlaps between the label points}

\item{mylevel}{A data frame corresponding to the pData}

\item{mylabsize}{A numeric value representing the police size to display for the different labels}

\item{dispelip}{A numeric value representing the ellipsoid dispersion}

\item{labeled}{A character to display labels and/or points}

\item{pal}{A color object from the RcolorBrewer package}
}
\value{
A ggplot object
}
\description{
PCAplot is a function that return a factoextra object of PCA type
}
