% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{PCAres}
\alias{PCAres}
\title{PCAres is a reactive function that computed a PCA of non-normalized data}
\usage{
PCAres(csvf, new_datapca)
}
\arguments{
\item{csvf}{a data frame corresponding to the WorkingSet}

\item{new_datapca}{a reactive data frame}
}
\value{
PCAres a reactive data frame with PCA attributes
}
\description{
PCAres is a reactive function that computed a PCA of non-normalized data
}
