% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/heatmtruncated.R
\name{plotHeatmaps}
\alias{plotHeatmaps}
\alias{plotHeatmaps}
\title{plotHeatmaps is a function that create a false color image with a dendogram added to the left side and the top}
\usage{
plotHeatmaps(exprData, groups, workingPath = getwd(), fileType = "png",
  cexcol = 1.5, cexrow = 1.5, colOrder = NULL, labrow = F,
  na.color = "black", scale = "row", rowv = list4, ColvOrd = list3,
  gpcol = list5, gpcolr = list6, distfunTRIX = list2,
  RowSideColor = c("gray25", "gray75"), palette.col = NULL, margins = c(8,
  8), my_palette = colorRampPalette(c("green", "black", "red"))(n = 75),
  mycex = 0.6, mypal = test, colid = NULL, showcol = T, showrow = F,
  notplot = F, geneSet = list7, genename = csvf, height = list8,
  rastering = myras)

plotHeatmaps(exprData, groups, workingPath = getwd(), fileType = "png",
  cexcol = 1.5, cexrow = 1.5, colOrder = NULL, labrow = F,
  na.color = "black", scale = "row", rowv = list4, ColvOrd = list3,
  gpcol = list5, gpcolr = list6, distfunTRIX = list2,
  RowSideColor = c("gray25", "gray75"), palette.col = NULL, margins = c(8,
  8), my_palette = colorRampPalette(c("green", "black", "red"))(n = 75),
  mycex = 0.6, mypal = test, colid = NULL, showcol = T, showrow = F,
  notplot = F, geneSet = list7, genename = csvf, height = list8,
  rastering = myras)
}
\arguments{
\item{exprData}{a data frame with all the individuals selected}

\item{groups}{a data frame with the corresponding groups}

\item{workingPath}{an access path}

\item{fileType}{a character value which extension corresponds to different file types}

\item{cexcol}{a positive numbers, used as cex.axis in for the row or column axis labeling}

\item{cexrow}{a positive numbers, used as cex.axis in for the row or column axis labeling}

\item{colOrder}{a positive numbers, used as cex.axis in for the row or column axis labeling}

\item{labrow}{a character vectors with row and column labels to use}

\item{na.color}{color to use for missing value}

\item{scale}{a character indicating if the values should be centered and scaled in either the row direction or the column direction, or none}

\item{rowv}{dendogram object}

\item{gpcol}{matrix with colors associated to each groups}

\item{gpcolr}{matrix with gray color depending on the clusters}

\item{distfunTRIX}{function that computes whether euclidian or pearson for Hierarchical Clustering}

\item{RowSideColor}{a character vector of length nrow containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{palette.col}{NULL}

\item{margins}{a numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively.}

\item{my_palette}{a character vetor of length 17}

\item{mycex}{a numeric value which aim is to change the size of the legend}

\item{mypal}{a character vetor of length 17}

\item{colid}{a character vectors with row and column labels to use; default NULL}

\item{showcol}{a boolean value used to hide or show the colnames}

\item{showrow}{a boolean value used to hide or show the rownames}

\item{notplot}{boolean calling or not the dev.off method}

\item{geneSet}{a data frame with the indexes corresponding to the sigificant genes}

\item{genename}{a data frame list corresponding to the gene names}

\item{k}{a numeric value which aim is to defined the treshold value to cut the dendogram}

\item{hclustGenes}{a function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms}

\item{meanGrp}{a boolean value to computes the mean for each groups; default = F}

\item{plotRowSideColor}{a boolean value that colors or not the rows side}

\item{Rowdistfun}{a function used to compute the distance for the rows}

\item{Coldistfun}{a function used to compute the distance for the columns}

\item{exprData}{data frame with all the individuals selected}

\item{geneSet}{data frame with the indexes corresponding to the sigificant genes}

\item{groups}{data frame with the corresponding groups}

\item{workingPath}{an access path}

\item{k}{numeric value which aim is to defined the treshold value to cut the dendogram}

\item{fileType}{character value which extension corresponds to different file types}

\item{cexcol}{positive numbers, used as cex.axis in for the row or column axis labeling}

\item{cexrow}{positive numbers, used as cex.axis in for the row or column axis labeling}

\item{colOrder}{positive numbers, used as cex.axis in for the row or column axis labeling}

\item{labrow}{character vectors with row and column labels to use}

\item{na.color}{color to use for missing value}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none}

\item{ColOrd}{positive numbers, used as cex.axis in for the row or column axis labeling}

\item{RowSideColor}{a character vector of length nrow containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{palette.col}{NULL}

\item{margins}{numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively.}

\item{my_palette}{a character vetor of length 17}

\item{mycex}{numeric value which aim is to change the size of the legend}

\item{mypal}{character vetor of length 17}

\item{colid}{character vectors with row and column labels to use; default NULL}

\item{showcol}{boolean value used to hide or show the colnames}

\item{showrow}{boolean value used to hide or show the rownames}

\item{genename}{data frame list corresponding to the gene names}
}
\value{
an heatmap object

data frame with the cluster and the corresponding genes
}
\description{
plotHeatmaps is a function that create a false color image with a dendogram added to the left side and the top

plotHeatmaps is a function that creates a false color image with a dendogram added to the left side and the top
}
