% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{plottopgenes}
\alias{plottopgenes}
\title{plottopgenes is an event reactive function which aim is to plot the top n genes selected by the user from the rendering data table}
\usage{
plottopgenes(topdegenes, venntopgenes, vennchoice, vennfinal, dispvenn)
}
\arguments{
\item{topdegenes}{clickable event button}

\item{venntopgenes}{numeric input}

\item{vennchoice}{reactive character vector}

\item{vennfinal}{a list of two data frames}

\item{dispvenn}{character input between probes and genes}
}
\value{
ggplot object
}
\description{
plottopgenes is an event reactive function which aim is to plot the top n genes selected by the user from the rendering data table
}
