% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{plottopgenesmean}
\alias{plottopgenesmean}
\title{plottopgenesmean is an event reactive function which aim is to plot the top n genes selected by the user from the rendering data table with the average logFC}
\usage{
plottopgenesmean(topdegenes, venntopgenes, vennchoice, vennfinal)
}
\arguments{
\item{topdegenes}{clickable event button}

\item{venntopgenes}{numeric input}

\item{vennchoice}{reactive character vector}

\item{vennfinal}{a list of two data frames}
}
\value{
ggplot object
}
\description{
plottopgenesmean is an event reactive function which aim is to plot the top n genes selected by the user from the rendering data table with the average logFC
}
