% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{probnamtoentrez}
\alias{probnamtoentrez}
\title{probnamtoentrez is a function which aim is to convert Gene symbols to entrez IDS}
\usage{
probnamtoentrez(davidres, mypack, gohm = F)
}
\arguments{
\item{davidres}{David Dataframe corresponding to the Functional Annotation Summary}

\item{mypack}{Annotation packages to get entrez ids from gene symbols}

\item{gohm}{A Boolean value to return a vector or lists of entrez ids}
}
\value{
A Vector or lists of entrez ids
}
\description{
probnamtoentrez is a function which aim is to convert Gene symbols to entrez IDS
}
