% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{probnamtoentrezvenn}
\alias{probnamtoentrezvenn}
\title{probnamtoentrezvenn is a function that convert Gene symbols to entrez IDS}
\usage{
probnamtoentrezvenn(venngenes, mypack)
}
\arguments{
\item{venngenes}{lists of genes}

\item{mypack}{package specific to the genome}
}
\value{
lists of entrez IDS
}
\description{
probnamtoentrezvenn is a function that convert Gene symbols to entrez IDS
}
