% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmodule.R
\name{queryDavid}
\alias{queryDavid}
\title{queryDavid is a shiny module which aims is to query the DWS for performing Functional Annotation or Gene Functional Classification}
\usage{
queryDavid(input, output, session, data, parent_session, tabsetpanid,
  tabPanel, hmana = F)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{data}{A reactive dataframe(s)}

\item{parent_session}{External session}

\item{tabsetpanid}{Widget tabset panel}

\item{tabPanel}{Plot tabset panel}

\item{hmana}{A boolen value (T for heatmap and F for Venn)}
}
\value{
A reactive statistical dataframe of S4 class
}
\description{
queryDavid is a shiny module which aims is to query the DWS for performing Functional Annotation or Gene Functional Classification
}
