% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcol.R
\name{renderncolour}
\alias{renderncolour}
\title{renderncolour is a shiny widget which aims is to return for each group a colorbox picker}
\usage{
renderncolour(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
renderncolour is a shiny widget which aims is to return for each group a colorbox picker
}
