% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnificentables.R
\name{renderoutputTable}
\alias{renderoutputTable}
\title{renderoutputTable is a shiny widget which aims is return a table in the gui}
\usage{
renderoutputTable(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
renderoutputTable is a shiny widget which aims is return a table in the gui
}
