% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{res.pca}
\alias{res.pca}
\title{res.pca is a function that computed a PCA of non-normalized data with the FactoMineR package}
\usage{
res.pca(workingset, scale = F)
}
\arguments{
\item{workingset}{A data frame corresponding to the WorkingSet}

\item{scale}{A boolean value; by default this value is set to False for non-normalized data}
}
\value{
A data frame with PCA attributes
}
\description{
res.pca is a function that computed a PCA of non-normalized data with the FactoMineR package
}
