% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{restabfc}
\alias{restabfc}
\title{This function returns a data frame of the element which are superior to a logFC cutoff (1.2,2,4,6 and 10) and for a defined pvalue}
\usage{
restabfc(alltop, pval, testrix, grepre)
}
\arguments{
\item{alltop}{A statistical dataframe}

\item{pval}{Cutoff pvalue}

\item{testrix}{A character value of the statistical criterions (pvalue, FDR)}

\item{grepre}{A list of subset dataframes}
}
\value{
A dataframe of the number of significant genes depending of both cutoff (logFC and pvalue)
}
\description{
This function returns a data frame of the element which are superior to a logFC cutoff (1.2,2,4,6 and 10) and for a defined pvalue
}
