% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{rowname}
\alias{rowname}
\title{rowname is a reactive function which aim is to hide or show the rownames}
\usage{
rowname(input)
}
\arguments{
\item{input$rowname}{a boolean radio button input}
}
\value{
rowname a reactive boolean value
}
\description{
rowname is a reactive function which aim is to hide or show the rownames
}
