% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{rownamtoX}
\alias{rownamtoX}
\title{rownametoX is a reactive function that change the rownames values}
\usage{
rownamtoX(csvf)
}
\arguments{
\item{csvf}{a data frame}
}
\value{
rownametoX a reactive data frame
}
\description{
rownametoX is a reactive function that change the rownames values
}
