% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{rowtoprob}
\alias{rowtoprob}
\title{rowtoprob is a function that return the probenames/transcripts for the corresponding indexes}
\usage{
rowtoprob(myven, pval, adj)
}
\arguments{
\item{myven}{A list of index for the different contrasts}

\item{pval}{A dataframe of the restable}

\item{adj}{A subset dataframe of the restable}
}
\value{
A list of transcripts and genes
}
\description{
rowtoprob is a function that return the probenames/transcripts for the corresponding indexes
}
