% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmodule.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis is a shiny widget action button for run the functional analysis}
\usage{
runAnalysis(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
runAnalysis is a shiny widget action button for run the functional analysis
}
