% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{Scree_plot}
\alias{Scree_plot}
\title{Scree_plot is a reactive function which aim is to display the eigenvalues of the data}
\usage{
Scree_plot(PCAres)
}
\arguments{
\item{PCAres}{a reactive data frame with PCA attributes}
}
\value{
\Screeplot a reactive plot
}
\description{
Scree_plot is a reactive function which aim is to display the eigenvalues of the data
}
