% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savemodules.R
\name{selFormat}
\alias{selFormat}
\title{selFormat is a shiny widget which aims is to specify the output format}
\usage{
selFormat(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
selFormat is a shiny widget which aims is to specify the output format
}
