% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmodule.R
\name{selSpecies}
\alias{selSpecies}
\title{selSpecies is a shiny widget which aims is to select the Species for the functional analysis}
\usage{
selSpecies(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Widget in the gui
}
\description{
selSpecies is a shiny widget which aims is to select the Species for the functional analysis
}
