% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{setvglobalvenn}
\alias{setvglobalvenn}
\title{setvglobalvenn is a function which aim is to return lists of each probes for the different set of intersections}
\usage{
setvglobalvenn(vennlist, adj)
}
\arguments{
\item{vennlist}{a list of genes for the different contrasts}

\item{adj}{dataframe subset of the alltoptable}
}
\value{
list of probes
}
\description{
setvglobalvenn is a function which aim is to return lists of each probes for the different set of intersections
}
