% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{Species}
\alias{Species}
\title{Species is a reactive function which aim is to return annotated packages for a specific genome}
\usage{
Species(Species, Speciesvenn)
}
\arguments{
\item{Species}{character input}

\item{Speciesvenn}{character input}
}
\description{
Species is a reactive function which aim is to return annotated packages for a specific genome
}
