% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnificentables.R
\name{stylishTables}
\alias{stylishTables}
\title{stylishTables is a shiny module which aims is to render a table in the gui}
\usage{
stylishTables(input, output, session, data, lengthpage = c("5", "10",
  "15"), extensions = c("Buttons", "Scroller"), scrollX = F,
  scrollY = F, pageLength = NULL, ordering = T, server = F,
  buttons = NULL, stateSave = NULL, searching = T, filter = NULL,
  dom = NULL, rownames = F, selection = "multiple",
  dupgenes = NULL, case = 1)
}
\arguments{
\item{input}{Internal}

\item{output}{Internal}

\item{session}{Internal}

\item{data}{A reactive dataframe}

\item{lengthpage}{A character vector (number of rows to display in the table)}

\item{scrollX}{A boolean value (horizontal scrolling )}

\item{scrollY}{A numeric value (Vertical scrolling)}

\item{pageLength}{A numeric value (number of rows in the output table)}

\item{ordering}{A boolean value}

\item{server}{A boolean value (server-side processing)}

\item{buttons}{A button collection that provides column visibility control}

\item{stateSave}{A boolean value (Enable or disable state saving)}

\item{searching}{A boolean value (Enable or disable research in table)}

\item{filter}{A function with two argumentsdataandparams}

\item{dom}{Define the table control elements to appear on the page and in what order.}

\item{rownames}{TRUE(show row names) or FALSE(hide row names) or a character vector of rownames;}

\item{selection}{the row/column selection mode (single or multiple selection or disable selection) when a table widget is rendered in a Shiny app}

\item{dupgenes}{A reactive character vector of the duplicated genes}

\item{case}{A Numeric value}

\item{extensionsa}{A character vector of the names of the DataTables extensions (check DT API)}
}
\value{
An outpout table in the gui
}
\description{
stylishTables is a shiny module which aims is to render a table in the gui
}
