% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{toJvenn}
\alias{toJvenn}
\title{toJvenn is a function which aims to convert a dataframe object to json format}
\usage{
toJvenn(myven, adj)
}
\arguments{
\item{myven}{A list of genes generates by the Vennlist function}

\item{adj}{A pvalue subset dataframe of the restable}
}
\value{
A json object
}
\description{
toJvenn is a function which aims to convert a dataframe object to json format
}
