% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{totalclust}
\alias{totalclust}
\title{totaclust is a reactive function which aim is to dynamically return a widget object of selectinput type ranging from 1 to the maximum number of cluster}
\usage{
totalclust(hmobj)
}
\arguments{
\item{hmobj}{data frame of the significant genes associated with the corresponding cluster index}
}
\value{
selectInput widget
}
\description{
totaclust is a reactive function which aim is to dynamically return a widget object of selectinput type ranging from 1 to the maximum number of cluster
}
