% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{totalvenn}
\alias{totalvenn}
\title{totalvenn is a function which aim is to return the total element of each interesections for the venn diagram}
\usage{
totalvenn(vennlist, adj)
}
\arguments{
\item{vennlist}{a list of genes for the different contrasts}

\item{adj}{dataframe subset of the alltoptable}
}
\value{
numeric value
}
\description{
totalvenn is a function which aim is to return the total element of each interesections for the venn diagram
}
