% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdegenes.R
\name{tracketCutoff}
\alias{tracketCutoff}
\title{cutoffElements is a global namespace containing a list of htmlOuput from widget inputs}
\usage{
tracketCutoff(id)
}
\arguments{
\item{id}{Shiny widget}
}
\value{
Widgets in the gui
}
\description{
cutoffElements is a global namespace containing a list of htmlOuput from widget inputs
}
