% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{transfheatm}
\alias{transfheatm}
\title{transfheatm is an observe function that change the global value after a clickable event}
\usage{
transfheatm(input, global)
}
\arguments{
\item{input}{heatm a numeric value by default set to 0}

\item{global}{a reactivevalues boolean value; by default set to FALSE}
}
\value{
global a boolean reactivalues set to TRUE
}
\description{
transfheatm is an observe function that change the global value after a clickable event
}
