% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{transform}
\alias{transform}
\title{transform is a function that transform a dataframe containing factor for different levels not optimal tho}
\usage{
transform(dataframe, toast)
}
\arguments{
\item{dataframe}{}
}
\value{

}
\description{
transform is a function that transform a dataframe containing factor for different levels not optimal tho
}
