% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Union}
\alias{Union}
\title{Union is a function that takes a list as argument and return an union of those lists}
\usage{
Union(x)
}
\arguments{
\item{x}{A list of elements}
}
\value{
A vector
}
\description{
Union is a function that takes a list as argument and return an union of those lists
}
\examples{
x <- c(sort(sample(1:20, 9)))
y <- c(sort(sample(3:23, 7)))
test = list()
test[[1]] = x
test[[2]] = y
Union(test) = 1  2  4  9 11 14 16 17 19  3  6 10 12 21
}
