% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{user_group}
\alias{user_group}
\title{user_group is a reactive function that summarise the significant genes depending on the pvalue with FC set to (1.2,2,4,6,10)#' Reactive function that return a list of data frame depending on the comparisons}
\usage{
user_group(adjusted, choix_test)
}
\arguments{
\item{adjusted}{list of three data frame corresponding to the grep of respectively Adj.pval, P.val and logFC columns}

\item{choix_test}{character corresponding to the defined contrast set by the user}
}
\value{
usergroup a reactive list containing three data frame for each contrast selected
}
\description{
user_group is a reactive function that summarise the significant genes depending on the pvalue with FC set to (1.2,2,4,6,10)#' Reactive function that return a list of data frame depending on the comparisons
}
