% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{vennchoice}
\alias{vennchoice}
\title{vennchoice is a reactive function that return user's selected comparisons}
\usage{
vennchoice(intscol)
}
\arguments{
\item{intscol}{character input}
}
\value{
character vector
}
\description{
vennchoice is a reactive function that return user's selected comparisons
}
