% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{Venncluster}
\alias{Venncluster}
\title{Venncluster is an event reactive function which aim is to interogate David web services database to collect relevant information about the list of genes for a specific intersection}
\usage{
Venncluster(GOvenn, vennfinal, SPecies, Speciesvenn)
}
\arguments{
\item{GOvenn}{clickable event button}

\item{vennfinal}{a list of two data frames}

\item{Speciesvenn}{character input}

\item{Species}{list of annotated elements}
}
\value{
david enrichment object
}
\description{
Venncluster is an event reactive function which aim is to interogate David web services database to collect relevant information about the list of genes for a specific intersection
}
