% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndiamgram.R
\name{Vennfinal}
\alias{Vennfinal}
\title{Vennfinal is a function which aim is to return an object containing a venn diagram}
\usage{
Vennfinal(myl, adj, cex = 1)
}
\arguments{
\item{myl}{a list of genes for the different contrasts}

\item{adj}{a data frame}

\item{cex}{a vector giving the size for each area label}
}
\value{
draw on the current device
}
\description{
Vennfinal is a function which aim is to return an object containing a venn diagram
}
