% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{venninter}
\alias{venninter}
\title{venninter is a reactive function which aim is to return a set of lists for each possible logical relations between a finite collection of different sets}
\usage{
venninter(vennlistshi, user_cont)
}
\arguments{
\item{vennlistshi}{list of probenames}

\item{user_cont}{character vector}
}
\value{
multiple lists
}
\description{
venninter is a reactive function which aim is to return a set of lists for each possible logical relations between a finite collection of different sets
}
