% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{vennlistshi}
\alias{vennlistshi}
\title{bool is a reactive function that return the bool value in the local environment}
\usage{
vennlistshi(csvf, user_cont, user_fc, regulation, pvalvenn, fcvenn)
}
\arguments{
\item{csvf}{a data frame}

\item{user_cont}{a subset data frame with the selected comparisons for the adj.p.val or p.val}

\item{user_fc}{a subset data frame with the selected comparisons for the logfc}

\item{regulation}{vector input}

\item{pvalvenn}{numeric input for the p value cutoff}

\item{fcvenn}{numeric input for the logfc value cutoff}
}
\value{
bool a reactive boolean outside the reactive environment

probven a reactive list of probenames
}
\description{
bool is a reactive function that return the bool value in the local environment
}
