% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{Vennplot}
\alias{Vennplot}
\title{Vennplot is a reactive function that return a plot object or a link if the user want to display more tha  5sets}
\usage{
Vennplot(Vennploted)
}
\arguments{
\item{Vennploted}{a reactive object}

\item{user_cont}{a reactive data frame with the selected contrast}

\item{input$vennsize}{the police size for the contrasts}

\item{vennlist}{a list}
}
\value{
Vennplot a reactive object to be plot

Vennplot is a reactive function that return an object of type venn if the number of set is stricly inferior to 6
or a link to a website if it's not

Vennploted a reactive object to be plot

bool a reactive boolean inside the reactive environment

output$bool is a reactive function that set the bool value to F

bool a reactive boolean inside the reactive environment
}
\description{
Vennplot is a reactive function that return a plot object or a link if the user want to display more tha  5sets
}
