% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Vennsev}
\alias{Vennsev}
\title{Vennsev is a function that is used for more than 5 intersections}
\usage{
Vennsev(myl, adj)
}
\arguments{
\item{myl}{a list of genes for the different contrasts}

\item{adj}{dataframe subset of the alltoptable}
}
\value{
plot device
}
\description{
Vennsev is a function that is used for more than 5 intersections
}
