% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{venntopgenes}
\alias{venntopgenes}
\title{venntopgenes is a reactive function which aim is to return the user's input top n genes}
\usage{
venntopgenes(topgenes)
}
\arguments{
\item{topgenes}{numeric input}
}
\value{
numeric input
}
\description{
venntopgenes is a reactive function which aim is to return the user's input top n genes
}
