% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{wclust}
\alias{wclust}
\title{wclust is a function that return a tabular file containing the top n genes for the different clusters, the go ids associated to this cluster, the id's term and the definition of the term}
\usage{
wclust(clusterlist, filename, min, top)
}
\arguments{
\item{clusterlist}{list of data frames}

\item{filename}{name of the output file}

\item{min}{GO ids that are not represented significally by default = 2}

\item{top}{top go ids to display}
}
\value{
txt file
}
\description{
wclust is a function that return a tabular file containing the top n genes for the different clusters, the go ids associated to this cluster, the id's term and the definition of the term
}
