% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{chartofa}
\alias{chartofa}
\title{This function returns a transformed data frame of character type to a data frame of factor type}
\usage{
chartofa(datach)
}
\arguments{
\item{datach}{A dataframe that associates samples to their respective biological conditions}
}
\value{
A dataframe of factor type
}
\description{
This function returns a transformed data frame of character type to a data frame of factor type
}
