% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{choix_test}
\alias{choix_test}
\title{choix_test is an eventreactive function in the aim of selecting different comparison after a clickable event}
\usage{
choix_test(input)
}
\arguments{
\item{input}{heatm a clickable button}
}
\value{
a reactive value of type character for the different comparisons selected
}
\description{
choix_test is an eventreactive function in the aim of selecting different comparison after a clickable event
}
