% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{col_choice1}
\alias{col_choice1}
\title{col_choice1 is a reactive function that return a character color}
\usage{
col_choice1(input)
}
\arguments{
\item{input}{col1 a character color for the highest values}
}
\value{
col_choice1  a reactive value
}
\description{
col_choice1 is a reactive function that return a character color
}
