% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{createdfsign}
\alias{createdfsign}
\title{Create a data frame containing the number of signficant genes for different conditions pval and log fc}
\usage{
createdfsign(adj)
}
\arguments{
\item{adj}{A subset dataframe}
}
\value{
A data frame
}
\description{
Create a data frame containing the number of signficant genes for different conditions pval and log fc
}
