% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{csvf}
\alias{csvf}
\title{Reactive function in the aim of loading csv files}
\usage{
csvf(input)
}
\arguments{
\item{input}{loaded files of csv format}
}
\value{
csvf a reactive value of type list containing three data frames toptable and workingset and the pData

apply the fread method for each element in the csvtest list

\csv a data frame object

Reactive function returned to the tab1.R

\showmark a reactive value of type boolean set to False
}
\description{
Reactive function in the aim of loading csv files
}
