% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutheatmap.R
\name{cutHeatmaps}
\alias{cutHeatmaps}
\title{cutHeatmaps if a function that takes as input an heatmap object and depending on the cut height and the cluster
choosen render a ggplot object or an heatmap object}
\usage{
cutHeatmaps(hmp, height, exprData, DEGres, groups, cexcol = 1, cexrow = 1,
  labrow = T, fileType = "png", scale = "row", meanGrp = F,
  col.hm = maPalette(low = "green", high = "red", mid = "black", k = 75),
  type = "None", las = 2, distfun = "cor", palette.col = NULL,
  num = 4, ...)
}
\arguments{
\item{hmp}{an heatmap object}

\item{height}{a numeric value to cut the dendogram}

\item{exprData}{a data frame with specific columns depending on the user's choices}

\item{DEGres}{a data frame corresponding to the xxx_topTableAll}

\item{groups}{a data frame of the choosen groups}

\item{cexcol}{a positive numbers, used as cex.axis in for the row or column axis labeling}

\item{cexrow}{a positive numbers, used as cex.axis in for the row or column axis labeling}

\item{labrow}{a character vectors with row and column labels to use}

\item{fileType}{a character to select the plot to display heatmap, boxplot or stripchart}

\item{scale}{a character indicating if the values should be centered and scaled in either the row direction or the column direction, or none}

\item{meanGrp}{a boolean value to computes the mean for each groups; default = F}

\item{col.hm}{a character vector}

\item{type}{a character to select the plot to display heatmap, boxplot or stripchart}

\item{las}{a numeric value}

\item{distfun}{function used to compute the distance (dissimilarity) between both rows and columns.}

\item{palette.col}{a character vector}

\item{num}{an item of the heatmap object corresponding to a specific cluster choosen by the user}

\item{...}{}
}
\value{
a ggplot object or heatmapply object
}
\description{
cutHeatmaps if a function that takes as input an heatmap object and depending on the cut height and the cluster
choosen render a ggplot object or an heatmap object
}
