% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{data_summary}
\alias{data_summary}
\title{data_summary is a reactive function that return the indexes for the signficant genes}
\usage{
data_summary(csvf, input)
}
\arguments{
\item{csvf}{data frame}

\item{input}{a numeric pvalue}
}
\value{
\datasummary a reactive data frame with the indexes corresponding to the sigificant genes for 5 Fold change 1.2,2,4,6,10
}
\description{
data_summary is a reactive function that return the indexes for the signficant genes
}
