% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{davidqueryvenn}
\alias{davidqueryvenn}
\title{davidqueryvenn is a function which aim is to querrying DWS and performing Functional Annotation Clustering}
\usage{
davidqueryvenn(entrezids, species)
}
\arguments{
\item{entrezids}{list of entrez IDS}

\item{species}{character name of the species whose genes are enriched}
}
\value{
david object
}
\description{
davidqueryvenn is a function which aim is to querrying DWS and performing Functional Annotation Clustering
}
