% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{distcor}
\alias{distcor}
\title{distcor is a function that computes the pearson correlation}
\usage{
distcor(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A distance matrix
}
\description{
distcor is a function that computes the pearson correlation
}
