% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{distcor}
\alias{distcor}
\title{distcor is a function that computes the distance correlation}
\usage{
distcor(x)
}
\arguments{
\item{x}{}
}
\value{
a matrix distance
}
\description{
distcor is a function that computes the distance correlation
}
