% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{distcos}
\alias{distcos}
\title{distcos is a function that computes the cosine similarity for a given matrix}
\usage{
distcos(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A distance matrix
}
\description{
distcos is a function that computes the cosine similarity for a given matrix
}
