% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/heatmtruncated.R
\name{disteucl}
\alias{disteucl}
\alias{disteucl}
\title{disteuc is a function that computes the euclidean correlation}
\usage{
disteucl(x)

disteucl(x)
}
\arguments{
\item{x}{}

\item{x}{}
}
\value{
a matrix distance

a matrix distance
}
\description{
disteuc is a function that computes the euclidean correlation

disteuc is a function that computes the euclidean distance matrices
}
