% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{disteucl}
\alias{disteucl}
\title{disteuc is a function that computes the euclidean distance matrices}
\usage{
disteucl(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A distance matrix
}
\description{
disteuc is a function that computes the euclidean distance matrices
}
