% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{eboulis}
\alias{eboulis}
\title{eboulis is a function which aim is to display the eigenvalues of the data with the package factoextra}
\usage{
eboulis(PCAres)
}
\arguments{
\item{PCAresa}{a data frame with PCA attributes}
}
\value{
a factoextra object
}
\description{
eboulis is a function which aim is to display the eigenvalues of the data with the package factoextra
}
