% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{entreztosymb}
\alias{entreztosymb}
\title{entreztosymb is a function which aim is to convert entrez IDS to gene symbols}
\usage{
entreztosymb(myentz, mypack)
}
\arguments{
\item{myentz}{lists of genes}

\item{mypack}{package specific to the genome}
}
\value{
lists of gene symbols
}
\description{
entreztosymb is a function which aim is to convert entrez IDS to gene symbols
}
