% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{evaluatesign}
\alias{evaluatesign}
\title{This function return a numeric value for the number of significant genes at a certain treshold}
\usage{
evaluatesign(adj, elem, pv)
}
\arguments{
\item{adj}{a data frame}

\item{elem}{a list}

\item{pv}{a list}
}
\value{
a data frame
}
\description{
This function return a numeric value for the number of significant genes at a certain treshold
}
