% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{evaluatesignpar}
\alias{evaluatesignpar}
\title{This function return an integer for the number of significant genes using parallelism}
\usage{
evaluatesignpar(adj, elem, pv)
}
\arguments{
\item{pv}{}
}
\value{
a data frame
}
\description{
This function return an integer for the number of significant genes using parallelism
}
