% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatingtables.R
\name{formating}
\alias{formating}
\title{formating is a alpha function of the higher elaborate decTestTRiX function}
\usage{
formating(adj, pval)
}
\arguments{
\item{adj}{A subset dataframe}

\item{pval}{Cutoff pvalue}
}
\value{
A numeric value of the number of significant genes
}
\description{
formating is a alpha function of the higher elaborate decTestTRiX function
}
