% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{formating}
\alias{formating}
\title{formating is a function alpha version of the higher elaborate decTestTRiX function}
\usage{
formating(adj, pval)
}
\arguments{
\item{pval}{}
}
\description{
formating is a function alpha version of the higher elaborate decTestTRiX function
}
