% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{global}
\alias{global}
\title{global is a ReactiveValues function}
\usage{
global(clicked)
}
\arguments{
\item{clicked}{bool set to FALSE}
}
\value{
clicked a boolean which can be TRUE or FALSE
}
\description{
global is a ReactiveValues function
}
