% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{hclustfun}
\alias{hclustfun}
\title{hclustfun is a function that performs a hiearchical cluster analysis with the method of Ward}
\usage{
hclustfun(d, e = "ward.D2")
}
\arguments{
\item{d}{A vector}

\item{e}{}
}
\value{
An object of class hclust
}
\description{
hclustfun is a function that performs a hiearchical cluster analysis with the method of Ward
}
