% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{hclustfun}
\alias{hclustfun}
\title{hclustfun is a function that performs a hiearchical cluster analysis with the Ward's method}
\usage{
hclustfun(d)
}
\arguments{
\item{d}{}
}
\value{
an object of class hclust
}
\description{
hclustfun is a function that performs a hiearchical cluster analysis with the Ward's method
}
