% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{heatmapfinal}
\alias{heatmapfinal}
\title{heatmapfinal is an isolate function that only react to a user's click on the heatmap button}
\usage{
heatmapfinal()
}
\arguments{
\item{new_data}{a data frame with all the individuals selected}

\item{formated}{a data frame with the indexes corresponding to the sigificant genes}

\item{new_group}{a data frame with the corresponding groups}

\item{workingPath}{the current user's repository}

\item{my_palette}{a vector of colors}

\item{k}{a numeric value which aim is to defined the treshold value to cut the dendogram input$clusters}

\item{Rowdistfun}{a character value set by the user to defined the method to calculate the dendogram matrix distance for the genes input$dist}

\item{Coldistfun}{a character value set by the user to defined the method to calculate the dendogram matrix distance for the contrasts input$dist}

\item{mycex}{a numeric value which aim is to change the size of the legend in the heatmap defined by the user input$legsize}

\item{cexrow}{a numeric value to change the size of the police legend for the rows input$rowsize}

\item{cexcol}{a numeric value to change the size of the police legend for the columns input$colsize}

\item{meanGrp}{a boolean value to compute or not the mean of each contrasts in the heatmap input$meangrp}

\item{mypal}{a list of values}

\item{showcol}{a boolean value used to hide or show the colnames input$colname}

\item{showrow}{a boolean value used to hide or show the rownames input$rowname}

\item{genename}{a data frame list corresponding to the gene names}
}
\value{
heatmapfinal a heatmap object
}
\description{
heatmapfinal is an isolate function that only react to a user's click on the heatmap button
}
