% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{heatmtoclust}
\alias{heatmtoclust}
\title{This function returns a data frame of the significant genes associated with the corresponding cluster index}
\usage{
heatmtoclust(hmp01_All, exprData, pval, height = 5)
}
\arguments{
\item{hmp01_All}{a heatmap object}

\item{exprData}{a vector of indices for the significant genes who have crossed the treshold pval and fc}

\item{pval}{a data frame of the alltoptable}

\item{height}{numeric value where the dendogram is cut off}
}
\value{
a data frame
}
\description{
This function returns a data frame of the significant genes associated with the corresponding cluster index
}
