% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{labeled}
\alias{labeled}
\title{labeled is a reactive function which aim is to display or not the labels in the PCA render plot}
\usage{
labeled(input)
}
\arguments{
\item{input$label}{a boolean}
}
\value{
\labeled a reactive  boolean depending of the user's choice to display or not the labels
}
\description{
labeled is a reactive function which aim is to display or not the labels in the PCA render plot
}
