% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{mycolgrppca}
\alias{mycolgrppca}
\title{mycolgrppca is a reactive function which aim is to display the total number of groups}
\usage{
mycolgrppca(csvf)
}
\arguments{
\item{csvf}{a dataframe}
}
\value{
mycolgrppca a reactive reorder dataframe
}
\description{
mycolgrppca is a reactive function which aim is to display the total number of groups
}
