% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formating.R
\name{myfinalfc}
\alias{myfinalfc}
\title{This function return a data frame of the element which are superior to a vector character 1.2,2,4,6 and 10 and for a defined pvalue}
\usage{
myfinalfc(alltop, pval)
}
\arguments{
\item{alltop}{a data frame}

\item{pval}{a numeric pvalue}
}
\value{
a data frame
}
\description{
This function return a data frame of the element which are superior to a vector character 1.2,2,4,6 and 10 and for a defined pvalue
}
