% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndiamgram.R
\name{myventocsv}
\alias{myventocsv}
\title{myventocsv is a function that create a csv file of the signficant genes for the different contrasts for a cutoff of 5%}
\usage{
myventocsv(myven, adj)
}
\arguments{
\item{myven}{a list of genes for the different contrasts}

\item{adj}{a data frame}
}
\value{

}
\description{
myventocsv is a function that create a csv file of the signficant genes for the different contrasts for a cutoff of 5%
}
