% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{new_data}
\alias{new_data}
\title{new_data is a reactive function that aim is to select specific individuals in the data frame}
\usage{
new_data(csvf, new_group)
}
\arguments{
\item{csvf}{Data frame corresponding to the Workingset}

\item{new_group}{a reactive factor with the corresponding groups selected}
}
\value{
new_data a reactive data frame with specific columns depending on the user's choices
}
\description{
new_data is a reactive function that aim is to select specific individuals in the data frame
}
