% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{new_group}
\alias{new_group}
\alias{new_group}
\title{list_ind is a reactive function in the aim of having selected groups in a list}
\usage{
new_group(csvf, choix_grp)

new_group(csvf, choix_grp)
}
\arguments{
\item{csvf}{a Data frame corresponding to the pData table}

\item{choix_grp}{a vector character corresponding to the defined groups set by the user}

\item{input}{specific of the individuals data frame}

\item{input}{heatm a clickable button}

\item{csvf}{a data frame of the pData}
}
\value{
a reactive list for the different individuals selected

new_group is an eventreactive function that select specific groups in the data frame

new_group an eventreactive factor with the corresponding groups selected

new_group a reactive new factor with the corresponding groups
}
\description{
list_ind is a reactive function in the aim of having selected groups in a list

new group is a reactive function that select specific groups in the data frame
}
