% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/heatmtruncated.R
\name{num2cols}
\alias{num2cols}
\alias{num2cols}
\title{num2cols is a function that gives a character vector of color names for a given numeric vector}
\usage{
num2cols(numVector, colp = palette())

num2cols(numVector, colp = palette())
}
\arguments{
\item{numVector}{a numeric vector}

\item{colp}{a character vector}

\item{numVector}{a numeric vector}

\item{colp}{a character vector}
}
\value{
a matrix object

a matrix object
}
\description{
num2cols is a function that gives a character vector of color names for a given numeric vector

num2cols is a function that gives a character vector of color names for a given numeric vector
}
