% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{p}
\alias{p}
\title{p is a reactive function that return a heatmap gplots object}
\usage{
p(input, heatmapfinal)
}
\arguments{
\item{input}{updateheatm a clickable button}

\item{heatmapfinal}{a function}
}
\value{
p heatmap object
}
\description{
p is a reactive function that return a heatmap gplots object
}
