% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R, R/myshiny.R
\name{PCAplot}
\alias{PCAplot}
\alias{PCAplot}
\title{PCAplot is a function that return a factoextra object of PCA type}
\usage{
PCAplot()

PCAplot()
}
\arguments{
\item{PCAres}{a data frame with PCA attributes}

\item{myax}{a numeric vector of length 2 specifying the dimensions to be plotted}

\item{elips}{a boolean value to add ellipse to the data distribution  for the different groups; default = False}

\item{rep}{a boolean value to avoid overlaps between the label points}

\item{mylevel}{a data frame corresponding to the pData}

\item{mylabsize}{a numeric value representing the police size to display for the different labels}

\item{dispelip}{a numeric value representing the ellipsoid dispersion}

\item{labeled}{a character to display labels and/or points}

\item{pal}{a color object from the RcolorBrewer package}

\item{brew.pal}{a color object from the RcolorBrewer package}

\item{PCAres}{a data frame with PCA attributes}

\item{label}{a boolean value depending of the user's choice to display or not the labels}

\item{habillage}{a data frame corresponding to the pData}

\item{addEllipseda}{a boolean value to add ellipse to the data distribution  for the different groups}

\item{ellipse.level}{a numeric value set to 0.8}

\item{repel}{a boolean value to avoid overlaps between the label points}

\item{axes}{a numeric vector of length 2 specifying the dimensions to be plotted}

\item{labelsize}{a numeric value representing the police size to display for the different labels}

\item{pointsize}{a numeric value representing the diameter of each points displayed in the graph}
}
\value{
a ggplot2 object

p a factoextra object
}
\description{
PCAplot is a function that return a factoextra object of PCA type

PCAplot is a function that return a factoextra object of PCA type
}
