% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{PCAres}
\alias{PCAres}
\title{PCAres is a reactive function that computed a PCA of non-normalized data}
\usage{
PCAres(csvf)
}
\arguments{
\item{csvf}{a data frame corresponding to the WorkingSet}
}
\value{
PCAres a reactive data frame with PCA attributes
}
\description{
PCAres is a reactive function that computed a PCA of non-normalized data
}
