% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{plotHeatmaps}
\alias{plotHeatmaps}
\title{plotHeatmaps is a function that creates a false color image (basically image(t(x))) with a dendrogram added to the left side and to the top.
Typically, reordering of the rows and columns according to some set of values (row or column means) within the restrictions imposed by the dendrogram is carried out.}
\usage{
plotHeatmaps(exprData, groups, cexcol = 1.5, cexrow = 1.5,
  colOrder = NULL, labrow = F, na.color = "black", scale = "row",
  rowv = list4, ColvOrd = list3, gpcol = list5, gpcolr = list6,
  distfunTRIX = list2, RowSideColor = c("gray25", "gray75"),
  palette.col = NULL, margins = c(8, 8),
  my_palette = colorRampPalette(c("green", "black", "red"))(n = 75),
  mycex = 0.6, mypal = test, colid = NULL, showcol = T,
  showrow = F, notplot = F, geneSet = list7, genename = csvf,
  height = list8, rastering = myras)
}
\arguments{
\item{exprData}{A data frame with all the individuals selected}

\item{groups}{A data frame with the corresponding groups}

\item{cexcol}{Positive numbers, used as cex.axis in for the row or column axis labeling}

\item{cexrow}{Positive numbers, used as cex.axis in for the row or column axis labeling}

\item{colOrder}{Positive numbers, used as cex.axis in for the row or column axis labeling}

\item{labrow}{A character vectors with row and column labels to use}

\item{na.color}{color to use for missing value}

\item{scale}{A character indicating if the values should be centered and scaled in either the row direction or the column direction, or none}

\item{rowv}{A dendogram object}

\item{ColvOrd}{Determines if and how the column dendrogram should be reordered}

\item{gpcol}{Matrix with colors associated to each groups}

\item{gpcolr}{Matrix with gray color depending on the clusters}

\item{distfunTRIX}{Function that computes whether euclidian or pearson for Hierarchical Clustering}

\item{RowSideColor}{A character vector of length nrow containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{palette.col}{NULL}

\item{margins}{A numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively.}

\item{my_palette}{A character vetor of length 17}

\item{mycex}{A numeric value which aim is to change the size of the legend}

\item{mypal}{A character vetor of length 17}

\item{colid}{A character vectors with row and column labels to use; default NULL}

\item{showcol}{A boolean value used to hide or show the colnames}

\item{showrow}{A boolean value used to hide or show the rownames}

\item{notplot}{A boolean value calling or not the dev.off method}

\item{geneSet}{A data frame with the indexes corresponding to the sigificant genes}

\item{genename}{A data frame list corresponding to the gene names}

\item{height}{A numeric value specifying where to cut the dendogram}

\item{rastering}{A boolean value to hide artefact pixels}
}
\value{
data frame with the cluster and the corresponding genes
}
\description{
plotHeatmaps is a function that creates a false color image (basically image(t(x))) with a dendrogram added to the left side and to the top.
Typically, reordering of the rows and columns according to some set of values (row or column means) within the restrictions imposed by the dendrogram is carried out.
}
