% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosearch.R
\name{probnamtoentrez}
\alias{probnamtoentrez}
\title{probnamtoentrez is a function that convert Gene symbols to entrez IDS}
\usage{
probnamtoentrez(hm01, mypack)
}
\arguments{
\item{hm01}{data frame}

\item{mypack}{package specific to the genome}
}
\value{
lists of entrez IDS
}
\description{
probnamtoentrez is a function that convert Gene symbols to entrez IDS
}
