% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{rowtoprob}
\alias{rowtoprob}
\title{rowtoprob is a function that return the probe names for the corresponding indexes}
\usage{
rowtoprob(myven, pval, adj)
}
\arguments{
\item{myven}{a list of index for the different contrasts}

\item{pval}{dataframe of the alltoptable}

\item{adj}{dataframe subset of the alltoptable}
}
\value{
list
}
\description{
rowtoprob is a function that return the probe names for the corresponding indexes
}
