% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Setdiff}
\alias{Setdiff}
\title{Setdiff is a function that remove the union of the y's from the common x's, x and y are lists of characters.}
\usage{
Setdiff(x, y)
}
\arguments{
\item{x}{List of characters}

\item{y}{List of characters}
}
\value{
A vector
}
\description{
Setdiff is a function that remove the union of the y's from the common x's, x and y are lists of characters.
}
