% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{setvglobalvenn}
\alias{setvglobalvenn}
\title{setvglobalvenn is a function which aim is to return lists of each probes for the different sets of the venn diagram}
\usage{
setvglobalvenn(vennlist, adj, dll = F)
}
\arguments{
\item{vennlist}{A list of genes for the different contrasts}

\item{adj}{A dataframe subset of the alltoptable}
}
\value{
A list of probes/transcripts for all the different sets with the name associated
}
\description{
setvglobalvenn is a function which aim is to return lists of each probes for the different sets of the venn diagram
}
