% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{topngenes}
\alias{topngenes}
\title{topngenes is a function to plot the top n genes for a defined intersection between comparison.s}
\usage{
topngenes(dfinter, mycont, inputtop, meandup = "probes", mean = F)
}
\arguments{
\item{dfinter}{list of intersection.s}

\item{mycont}{character Vector}

\item{inputtop}{numeric value}

\item{meandup}{character}
}
\value{
ggplot2 barplot
}
\description{
topngenes is a function to plot the top n genes for a defined intersection between comparison.s
}
