% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmtruncated.R
\name{truncatedhat}
\alias{truncatedhat}
\title{hmbis is an event reactive function that pre-computed hierarchical clustering on microarray data}
\usage{
truncatedhat(exprData, geneSet, groups, workingPath = getwd(), k = 3,
  fileType = "png", colOrder = NULL, na.color = "black",
  hclustGenes = T, meanGrp = F, plotRowSideColor = T, mypal = NULL,
  RowSideColor = c("gray25", "gray75"), Rowdistfun = "correlation",
  Coldistfun = "correlation", palette.col = NULL, notplot = T,
  genename = pval)
}
\arguments{
\item{exprData}{data frame with all the individuals selected}

\item{geneSet}{data frame with the indexes corresponding to the sigificant genes}

\item{groups}{data frame with the corresponding groups}

\item{workingPath}{access path}

\item{k}{numeric value which aim is to defined the treshold value to cut the dendogram}

\item{fileType}{character value which extension corresponds to different file types}

\item{colOrder}{positive numbers, used as cex.axis in for the row or column axis labeling}

\item{na.color}{color used for missing values}

\item{hclustGenes}{function used to compute the hierarchical clustering when Rowv or Colv are not dendrograms}

\item{meanGrp}{boolean value to computes the mean for each groups; default = F}

\item{plotRowSideColor}{boolean value that colors or not the rows side}

\item{RowSideColor}{character vector of length nrow containing the color names for a vertical side bar that may be used to annotate the rows of x.}

\item{Rowdistfun}{function used to compute the distance for the rows}

\item{Coldistfun}{function used to compute the distance for the columns}

\item{palette.col}{NULL}

\item{notplot}{boolean}
}
\value{
list of objects which aim is to being passed as argument in the plotHeatmaps function
}
\description{
hmbis is an event reactive function that pre-computed hierarchical clustering on microarray data
}
