% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myshiny.R
\name{user_cont}
\alias{user_cont}
\title{user_cont is a reactive function that  return the contrast selected by the user}
\usage{
user_cont(adjusted, choix_cont)
}
\arguments{
\item{adjusted}{a data frame corresponding to the contrasts selected}

\item{choix_cont}{a set of character}
}
\value{
user_cont a reactive data frame with the contrast selected
}
\description{
user_cont is a reactive function that  return the contrast selected by the user
}
