% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Vennfinal}
\alias{Vennfinal}
\title{Vennfinal is a function which aim is to return an object containing a venn diagram}
\usage{
Vennfinal(myl, adj, cex = 1, cutoffpval, cutofffc, statimet,
  meandup = "probes", pval)
}
\arguments{
\item{myl}{a list of genes for the different contrasts}

\item{adj}{dataframe subset of the alltoptable}

\item{cex}{vector giving the size for each area label (length = 1/3/7/15 based on set-number)}

\item{cutoffpval}{numeric value}

\item{cutofffc}{numeric value}

\item{statimet}{character}

\item{meandup}{character}

\item{pval}{data frame of the alltoptable}
}
\value{
final draw on the current device of the venn diagram
}
\description{
Vennfinal is a function which aim is to return an object containing a venn diagram
}
