% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{vennfinal}
\alias{vennfinal}
\title{vennfinal is a reactive function which return a list of data frame corresponding to the computationnal mean of each logFC for the possible logical relations between a finite collection of different sets
and a data frame with as primary key the probenames associated with the corresponding gene names and logFC}
\usage{
vennfinal(vennchoice, adjusted, dispvenn, venninter)
}
\arguments{
\item{vennchoice}{reactive character vector}

\item{adjusted}{dataframe subset of the alltoptable}

\item{dispvenn}{character input between probes and genes}

\item{venninter}{multiple lists of probenames}
}
\value{
a list of two data frames
}
\description{
vennfinal is a reactive function which return a list of data frame corresponding to the computationnal mean of each logFC for the possible logical relations between a finite collection of different sets
and a data frame with as primary key the probenames associated with the corresponding gene names and logFC
}
