% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Vennfinal}
\alias{Vennfinal}
\title{Vennfinal is a function which aim is to return an object containing a venn diagram (old function with Venndiagram had been change with Jvenn)}
\usage{
Vennfinal(myl, adj, cex = 1, cutoffpval, cutofffc, statimet,
  meandup = "probes", pval, mycol = "")
}
\arguments{
\item{myl}{A list of genes for the different contrasts}

\item{adj}{A dataframe subset of the alltoptable}

\item{cex}{A vector giving the size for each area label (length = 1/3/7/15 based on set-number)}

\item{cutoffpval}{Cut-off for pvalue}

\item{cutofffc}{Cut-off for absolute log2 fold-change}

\item{statimet}{A character}

\item{meandup}{A character}

\item{pval}{A data frame of the restable}

\item{mycol}{A character vector of selected contrast(s)}
}
\value{
Draw on the current device a venn diagram
}
\description{
Vennfinal is a function which aim is to return an object containing a venn diagram (old function with Venndiagram had been change with Jvenn)
}
