% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Venndiamgram.R
\name{Vennlist}
\alias{Vennlist}
\title{Vennlist is a function which aim is to return a list of signficant genes for a treshold pvalue of 5%}
\usage{
Vennlist(pval, adj)
}
\arguments{
\item{pval}{a data frame}

\item{adj}{a data frame with the contrast selected}
}
\value{
a list
}
\description{
Vennlist is a function which aim is to return a list of signficant genes for a treshold pvalue of 5%
}
