% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennplot.R
\name{Vennlist}
\alias{Vennlist}
\title{Vennlist is a function which aim is to return a list of signficant genes for a treshold defined by the user}
\usage{
Vennlist(adj, fc, regulation, cutoffpval, cutofffc)
}
\arguments{
\item{adj}{dataframe subset of the alltoptable}

\item{fc}{dataframe subset of the alltoptable}

\item{regulation}{character for up both or down}

\item{cutoffpval}{numeric value}

\item{cutofffc}{numeric value}
}
\value{
list.s
}
\description{
Vennlist is a function which aim is to return a list of signficant genes for a treshold defined by the user
}
