% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final.R
\name{Vennplot}
\alias{Vennplot}
\title{Vennplot is a reactive function that return a plot object or a link if the user want to display more tha  5sets}
\usage{
Vennplot(Vennploted)
}
\arguments{
\item{Vennploted}{a reactive object}

\item{user_cont}{a subset data frame with the selected comparisons for the adj.p.val or p.val}

\item{vennsize}{the police size for the contrasts}

\item{vennlistshi}{a list of probenames}

\item{pvalvenn}{numeric input for the p value cutoff}

\item{fcvenn}{numeric input for the logfc value cutoff}

\item{methodforvenn}{character input}

\item{dispvenn}{character input for plot a venn diagram with probes or genes}

\item{csvf}{data frame corresponding to the alltoptable}
}
\value{
Vennplot a reactive object to be plot

Vennploted a reactive object to be plot

bool a reactive boolean inside the reactive environment

bool a reactive boolean inside the reactive environment
}
\description{
Vennplot is a reactive function that return a plot object or a link if the user want to display more tha  5sets
}
